refactor: BaseEntity 패키지 global로 수정
