+++ backend/src/main/java/woowacourse/touroot/member/domain/Member.java
@@ -9,7 +9,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import woowacourse.touroot.entity.BaseEntity;
+import woowacourse.touroot.global.entity.BaseEntity;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PUBLIC)
