+++ backend/src/main/java/woowacourse/touroot/place/domain/Place.java
@@ -1,11 +1,15 @@
 package woowacourse.touroot.place.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import woowacourse.touroot.entity.BaseEntity;
+import woowacourse.touroot.global.entity.BaseEntity;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -31,7 +35,7 @@ public class Place extends BaseEntity {
     public Place(String name, String latitude, String longitude, String googlePlaceId) {
         this(null, name, latitude, longitude, googlePlaceId);
     }
-    
+
     public Place(String name, String latitude, String longitude) {
         this(null, name, latitude, longitude, null);
     }
