+++ backend/src/main/java/woowacourse/touroot/travelogue/domain/TraveloguePhoto.java
@@ -13,7 +13,7 @@ import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import woowacourse.touroot.entity.BaseEntity;
+import woowacourse.touroot.global.entity.BaseEntity;
 
 @Getter
 @EqualsAndHashCode(of = "id", callSuper = false)
